refactor: use shared component for fertilizer icons
