+++ .changeset/twenty-cobras-wish.md
@@ -1,0 +1,5 @@
+---
+"@nmi-agro/fdm-app": patch
+---
+
+Use shared component for fertilizer icons
