+++ fdm-app/app/components/blocks/rotation/fertilizer-display.tsx
@@ -1,7 +1,7 @@
 import type { Row } from "@tanstack/react-table"
-import { Circle, Diamond, Square, Triangle } from "lucide-react"
 import React from "react"
 import { NavLink } from "react-router"
+import { FertilizerIcon } from "@/app/components/custom/fertilizer-icon"
 import { Badge } from "~/components/ui/badge"
 import { Tooltip, TooltipContent, TooltipTrigger } from "~/components/ui/tooltip"
 import type { FieldRow, RotationExtended } from "./columns"
@@ -9,37 +9,6 @@ import type { FieldRow, RotationExtended } from "./columns"
 type FertilizerDisplayProps = {
   row: Row<RotationExtended>
 }
-
-const fertilizerIconClassMap = {
-  manure: {
-    text: "text-yellow-600",
-    fill: {
-      "600": "fill-yellow-600",
-      "300": "fill-yellow-300",
-    },
-  },
-  mineral: {
-    text: "text-sky-600",
-    fill: {
-      "600": "fill-sky-600",
-      "300": "fill-sky-300",
-    },
-  },
-  compost: {
-    text: "text-green-600",
-    fill: {
-      "600": "fill-green-600",
-      "300": "fill-green-300",
-    },
-  },
-  other: {
-    text: "text-gray-600",
-    fill: {
-      "600": "fill-gray-600",
-      "300": "fill-gray-300",
-    },
-  },
-} as const
 
 export const FertilizerDisplay: React.FC<FertilizerDisplayProps> = ({ row }) => {
   const fertilizerDisplay = React.useMemo(() => {
@@ -58,9 +27,6 @@ export const FertilizerDisplay: React.FC<FertilizerDisplayProps> = ({ row }) => 
           const isFertilizerUsedOnAllFieldsForThisCultivation =
             row.original.type === "field" ||
             fields.every((field) => field.fertilizers.some((f) => f.p_id === fertilizer.p_id))
-          const fertilizerIconFillShade = isFertilizerUsedOnAllFieldsForThisCultivation
-            ? "600"
-            : "300"
 
           const component = (
             <NavLink
@@ -69,23 +35,10 @@ export const FertilizerDisplay: React.FC<FertilizerDisplayProps> = ({ row }) => 
             >
               <Badge variant="outline" className="text-muted-foreground gap-1">
                 <span>
-                  {fertilizer.p_type === "manure" ? (
-                    <Square
-                      className={`size-3 ${fertilizerIconClassMap.manure.text} ${fertilizerIconClassMap.manure.fill[fertilizerIconFillShade]}`}
-                    />
-                  ) : fertilizer.p_type === "mineral" ? (
-                    <Circle
-                      className={`size-3 ${fertilizerIconClassMap.mineral.text} ${fertilizerIconClassMap.mineral.fill[fertilizerIconFillShade]}`}
-                    />
-                  ) : fertilizer.p_type === "compost" ? (
-                    <Triangle
-                      className={`size-3 ${fertilizerIconClassMap.compost.text} ${fertilizerIconClassMap.compost.fill[fertilizerIconFillShade]}`}
-                    />
-                  ) : (
-                    <Diamond
-                      className={`size-3 ${fertilizerIconClassMap.other.text} ${fertilizerIconClassMap.other.fill[fertilizerIconFillShade]}`}
-                    />
-                  )}
+                  <FertilizerIcon
+                    dimmed={!isFertilizerUsedOnAllFieldsForThisCultivation}
+                    p_type={fertilizer.p_type ?? "other"}
+                  />
                 </span>
                 {fertilizer.p_name_nl}
               </Badge>
